codeBin/requirements.txt
Django==1.11.5
djangorestframework==3.6.4
MySQL-python==1.2.5
Pygments==2.2.0
pytz==2017.2
